+++ src/components/tab/component.jsx
@@ -1,0 +1,5 @@
+export const Tab = ({ title, isActive, onClick }) => {
+  return <button disabled={isActive} onClick={onClick}>
+    {title}
+  </button>
+}
